Organization and fixes
- New general utils file on CUD utils proyect
- Moved some utils function to general_utils.h
- Clean-up and fixes on macro_tools.h
- New experimental extended_base and number(big numbers) classes
- Moved GCD function from algorithm.h to arithmetic.h
- Added a way to specify your own container type for vector class
- Changed requirements con Complex concept

template_traits.h:
- New max bytes size conditional
- New conditional container base on size
- Added alias for some specific cases

+++ DTMLib/src/utilities/mutils.h
@@ -40,7 +40,7 @@ namespace duc::math_utils {
 /// \Warn	This function modify its parameters.
 /// \Return Returns the $base multiply by itself $exponent times.
 /// 
-	constexpr double powerPositiveInteger(uint64_t base, uint64_t exponent) noexcept {
+	constexpr double powerPositiveInteger(double base, uint64_t exponent) noexcept {
 		double result = 1;
 
 		while (exponent > 0) {
